Don't set max_players in the game host provider on init
There currently is no way to override this value, so it just results in confusion on the side of the reporting systems since the number of current players can be greater than the number of max players. For now this leaves the GHA entirely responsible for managing the max_players number to the best of its ability.