Ensure post init funciton is called if netimgui connection is attempted.
Split global and individual instance initialization.  This was hidden in a lot of cases for normal clients, but caused issues in PIE.

Apply docking flag every frame, rather than once on init.  This is to prevent multi-context cases like PIE only setting the flag on one context.  For PIE specifically, it was setting it on the editor's context, not the PIE window contexts.