Use join time as a timebreaker and only consider joined players when determining scout for a platform session.
This should make it the selection more stable, where the first player to accept can go ahead and form the session immediately upon getting a response without being contested in the extreme majority of cases.

This is to work around a bug where platform information could not be determined until a client joins, making the selection less stable than intended.  It was originally intended to consider all potential players so that players accepting massi invites (ex: from matchmaking) would not fight over who is the scout.